+++ pymrio/core/mriosystem.py
@@ -873,16 +873,15 @@ class Extension(CoreSystem):
             Row vector with population per region
         """
 
+        # TODO: This should only be used for calculating the full system.
+        # TODO There needs to be a new method for calculating the system for a different demand vector
+
         if Y_agg is None:
             try:
                 Y_agg = Y.groupby(level="region", axis=1, sort=False).sum()
 
             except (AssertionError, KeyError):
-                Y_agg = Y.groupby(
-                    level=0,
-                    axis=1,
-                    sort=False
-                ).sum()
+                Y_agg = Y.groupby(level=0, axis=1, sort=False).sum()
 
         y_vec = Y.sum(axis=0)
 
@@ -939,7 +938,7 @@ class Extension(CoreSystem):
                 return
             else:
                 self.D_cba, self.D_pba, self.D_imp, self.D_exp = calc_accounts(
-                    self.S, L, Y_agg, self.get_sectors().size
+                    self.S, L, Y_agg
                 )
                 logging.debug("{} - Accounts D calculated".format(self.name))
 
@@ -957,8 +956,7 @@ class Extension(CoreSystem):
                 )
             except (AssertionError, KeyError):
                 self.D_cba_reg = (
-                    self.D_cba.groupby(level=0, axis=1, sort=False).sum()
-                    + F_Y_agg
+                    self.D_cba.groupby(level=0, axis=1, sort=False).sum() + F_Y_agg
                 )
             try:
                 self.D_pba_reg = (
@@ -967,15 +965,18 @@ class Extension(CoreSystem):
                 )
             except (AssertionError, KeyError):
                 self.D_pba_reg = (
-                    self.D_pba.groupby(level=0, axis=1, sort=False).sum()
-                    + F_Y_agg
+                    self.D_pba.groupby(level=0, axis=1, sort=False).sum() + F_Y_agg
                 )
             try:
-                self.D_imp_reg = self.D_imp.groupby(level="region", axis=1, sort=False).sum()
+                self.D_imp_reg = self.D_imp.groupby(
+                    level="region", axis=1, sort=False
+                ).sum()
             except (AssertionError, KeyError):
                 self.D_imp_reg = self.D_imp.groupby(level=0, axis=1, sort=False).sum()
             try:
-                self.D_exp_reg = self.D_exp.groupby(level="region", axis=1, sort=False).sum()
+                self.D_exp_reg = self.D_exp.groupby(
+                    level="region", axis=1, sort=False
+                ).sum()
             except (AssertionError, KeyError):
                 self.D_exp_reg = self.D_exp.groupby(level=0, axis=1, sort=False).sum()
 
@@ -1585,7 +1586,7 @@ class Extension(CoreSystem):
                 factors_cleaned_gathered.append(fac_rest)
 
         for imissi in impacts_stressors_missing:
-            logging.warn(
+            logging.warning(
                 f"Impact >{imissi}< removed - calculation requires stressors "
                 f"not present in extension >{self.name}<"
             )